Fix for quote within quoted fields issue.
Consider the following field to be parsed in non-strict mode, with
standard eol and separator:

"\"Hello\" \"Hello2\" \"Hello3\""

I.e., besides the starting and closing quotes, there are some additional
(invalid!) quotes in the input field.
Without the fix, the following result is received:

" \"Hello2\" \"Hello3\""

This is not consistent: the first part is always lost.

After this fix, the result will be:

"Hello\" \"Hello2\" \"Hello3"

That is, the opening and closing quotes are treated as "normal" ones,
whilest those within the field are part of the value. (The last quote
should be immediately followed by a separater, new line, or EOF, to be
recognized as such.)